backend_backup/requirements.txt
fastapi>=0.103.1
uvicorn>=0.22.0
httpx>=0.24.1
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pytest>=7.4.0
# Specific version to avoid the import error with _QUERY_OPTIONS
pymongo==4.5.0
motor>=3.3.1,<4.0.0
apscheduler>=3.10.1
python-telegram-bot==20.8
pytest-asyncio>=0.21.1
lxml>=4.9.3
pydantic>=2.8.0
python-multipart>=0.0.5
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0
requests==2.31.0
aiohttp==3.8.5
bcrypt==4.0.1
selenium==4.15.2
webdriver-manager==4.0.1
bs4==0.0.1
sqlalchemy==2.0.27
PyPDF2==3.0.1
python-docx==1.0.1
google-ads==22.1.0
psycopg2-binary==2.9.9
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.105.0 